cambios en streamlit, quedan ultimos retoques en about us y modificar el codigo de la pagina our code + autor original
