+++ data/video_sample/IMG_5693.mp4
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d0c599792fabb0817d17a8b06f73e661c819a4b4671c8a818e522b0d76c5f313
+size 151083952
